ai-menu-parser/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
